+++ assets/services.js
@@ -2,7 +2,7 @@
 (function () {
   "use strict";
 
-  var version = "85.0.0";
+  var version = "86.0.0";
   var expectedServiceCount = 24;
 
   var manifest = [